archive_2025/abstraction_2025/pyproject.toml
[project]
name = "abstraction_2025"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
]
